Split Class directory and changed to multiple with more useful names. Removed remaining magic numbers and strings from, enemies, gamestates, collisionManager, and projectiles
