+++ Sprint0/Configs/GameConfig.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace sprint0.Classes
+namespace sprint0.Configs
 {
     public class GameConfig
     {
