+++ Sprint0/Enemies/AquamentusBoss.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using Microsoft.Xna.Framework;
-using sprint0.Classes;
 using sprint0.Interfaces;
 using sprint0.Factories;
