+++ Sprint0/Enemies/GoriyaStateMachine.cs
@@ -4,7 +4,6 @@ using Microsoft.Xna.Framework;
 using sprint0.Factories;
 using sprint0.Projectiles;
 using sprint0.Interfaces;
-using sprint0.RoomClasses;
 
 namespace sprint0.Enemies
 {
