+++ Sprint0/FileReaderClasses/LevelFileReader.cs
@@ -4,7 +4,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Xml;
 using Microsoft.Xna.Framework;
-using sprint0.Classes;
+using sprint0.Configs;
 using sprint0.FileReaderClasses;
 
 namespace sprint0.FileReaderClasses
