+++ Sprint0/GameStateClasses/GameWinState.cs
@@ -3,9 +3,7 @@
 namespace sprint0.GameStateClasses;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Audio;
-using sprint0.Sprites;
 using sprint0.Factories;
-using static System.Formats.Asn1.AsnWriter;
 
 public class GameWinState : AGameState
 {
