+++ Sprint0/ItemClasses/Pickups/BombPickup.cs
@@ -1,6 +1,5 @@
 using System;
 using Microsoft.Xna.Framework;
-using sprint0.Classes;
 using sprint0.Factories;
 using sprint0.Interfaces;
