+++ Sprint0/ItemClasses/Pickups/KeyPickup.cs
@@ -1,5 +1,4 @@
 using Microsoft.Xna.Framework;
-using sprint0.Classes;
 using sprint0.Factories;
 using sprint0.Interfaces;
