+++ Sprint0/PlayerClasses/Abilities/PlayerAbilityManager.cs
@@ -1,6 +1,5 @@
 using System.Collections.Generic;
 using Microsoft.Xna.Framework;
-using sprint0.Classes;
 
 namespace sprint0.PlayerClasses.Abilities;
