+++ Sprint0/PlayerClasses/DamagedPlayer.cs
@@ -1,12 +1,11 @@
 using System;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using sprint0.Classes;
 using sprint0.Factories;
 using sprint0.Interfaces;
 using sprint0.PlayerClasses.Abilities;
 
-namespace sprint0.PlayerClasses; 
+namespace sprint0.PlayerClasses;
 
 public class DamagedPlayer : IPlayer {
