+++ Sprint0/PlayerClasses/Player.cs
@@ -6,9 +6,8 @@ using sprint0.PlayerClasses.Abilities;
 using sprint0.RoomClasses;
 using sprint0.Sound;
 using System;
-using sprint0.Classes;
 
-namespace sprint0.PlayerClasses; 
+namespace sprint0.PlayerClasses;
 
 public class Player : IPlayer {
