+++ Sprint0/PlayerClasses/PlayerSwordDownState.cs
@@ -1,9 +1,8 @@
-using sprint0.Classes;
 using sprint0.Factories;
 using sprint0.Interfaces;
 
 
-namespace sprint0.PlayerClasses; 
+namespace sprint0.PlayerClasses;
 
 public class PlayerSwordDownState : PlayerSwordState {
