+++ Sprint0/PlayerClasses/PlayerSwordLeftState.cs
@@ -1,11 +1,10 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using sprint0.Classes;
 using sprint0.Factories;
 using sprint0.Interfaces;
 
 
-namespace sprint0.PlayerClasses; 
+namespace sprint0.PlayerClasses;
 
 public class PlayerSwordLeftState : PlayerSwordState {
