+++ Sprint0/PlayerClasses/PlayerSwordRightState.cs
@@ -1,8 +1,7 @@
-using sprint0.Classes;
 using sprint0.Factories;
 using sprint0.Interfaces;
 
-namespace sprint0.PlayerClasses; 
+namespace sprint0.PlayerClasses;
 
 public class PlayerSwordRightState : PlayerSwordState {
     public PlayerSwordRightState(Player player) {
